# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flipper AI API
  version: 0.1.0
  description: |
    Multi-marketplace flipping tool API. Find undervalued items, analyze flip potential
    with AI, manage opportunities, communicate with sellers, and list items for resale.
  contact:
    name: Axovia AI
    email: axoviaai@gmail.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://flipper-ai.vercel.app/api
    description: Production
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
    Platform:
      type: string
      enum: [CRAIGSLIST, FACEBOOK_MARKETPLACE, EBAY, OFFERUP, MERCARI]
...
    PostingStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, POSTED, FAILED, CANCELLED]
...
    SearchConfig:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        platform: { $ref: '#/components/schemas/Platform' }
        location: { type: string }
        category: { type: string, nullable: true }
        keywords: { type: string, nullable: true }
        minPrice: { type: number, nullable: true }
        maxPrice: { type: number, nullable: true }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    ScraperJob:
      type: object
      properties:
        id: { type: string }
        platform: { $ref: '#/components/schemas/Platform' }
        location: { type: string, nullable: true }
        category: { type: string, nullable: true }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    PostingQueueItem:
      type: object
      properties:
        id: { type: string }
        listingId: { type: string }
        targetPlatform: { $ref: '#/components/schemas/Platform' }
        status: { $ref: '#/components/schemas/PostingStatus' }
        askingPrice: { type: number, nullable: true }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        scheduledAt: { type: string, format: date-time, nullable: true }
        listing:
          type: object
          properties:
            id: { type: string }
            title: { type: string }
            platform: { $ref: '#/components/schemas/Platform' }
            askingPrice: { type: number }
            imageUrls: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
security:
  - session: []
paths:
...
  /search-configs:
    get:
      tags: [Search Configs]
      summary: List search configurations
      parameters:
        - { name: enabled, in: query, schema: { type: string, enum: ['true', 'false'] } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs: { type: array, items: { $ref: '#/components/schemas/SearchConfig' } }
                  total: { type: integer }
    post:
      tags: [Search Configs]
      summary: Create a search configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, platform, location]
              properties:
                name: { type: string, maxLength: 200 }
                platform: { $ref: '#/components/schemas/Platform' }
                location: { type: string, maxLength: 500 }
                category: { type: string, maxLength: 200 }
                keywords: { type: string, maxLength: 1000 }
                minPrice: { type: number, minimum: 0 }
                maxPrice: { type: number, minimum: 0 }
                enabled: { type: boolean, default: true }
      responses:
        '201': { description: Created }
...
  /scraper-jobs:
    get:
      tags: [Scraper Jobs]
      summary: List scraper jobs
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: platform, in: query, schema: { $ref: '#/components/schemas/Platform' } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: '#/components/schemas/ScraperJob' } }
                  total: { type: integer }
    post:
      tags: [Scraper Jobs]
      summary: Create a scraper job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform: { $ref: '#/components/schemas/Platform' }
                location: { type: string, maxLength: 500 }
                category: { type: string, maxLength: 200 }
      responses:
        '201': { description: Created }
...
  /posting-queue:
    get:
      tags: [Posting Queue]
      summary: List posting queue items
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PostingStatus' } }
        - { name: targetPlatform, in: query, schema: { $ref: '#/components/schemas/Platform' } }
        - { name: listingId, in: query, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PostingQueueItem' } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [Posting Queue]
      summary: Queue listing for resale (single or batch)
      description: |
        Supports two modes:
        - **Single**: `{ listingId, targetPlatform, ... }`
        - **Batch**: `{ listingId, platforms: [...], ... }` to post to multiple marketplaces at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [listingId, targetPlatform]
                  properties:
                    listingId: { type: string }
                    targetPlatform: { $ref: '#/components/schemas/Platform' }
                    askingPrice: { type: number }
                    title: { type: string, maxLength: 200 }
                    description: { type: string, maxLength: 5000 }
                    scheduledAt: { type: string, format: date-time }
                - type: object
                  required: [listingId, platforms]
                  properties:
                    listingId: { type: string }
                    platforms: { type: array, items: { $ref: '#/components/schemas/Platform' }, minItems: 1, maxItems: 5 }
                    askingPrice: { type: number }
                    title: { type: string, maxLength: 200 }
                    description: { type: string, maxLength: 5000 }
      responses:
        '201': { description: Queued }
...
  /user/settings:
    get:
      tags: [User Settings]
      summary: Get user settings
      responses:
        '200':
          description: User settings (API keys masked)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      userId: { type: string }
                      hasOpenaiApiKey: { type: boolean }
                      llmModel: { type: string }
                      discountThreshold: { type: number }
                      autoAnalyze: { type: boolean }
    patch:
      tags: [User Settings]
      summary: Update user settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                openaiApiKey: { type: string }
                llmModel: { type: string }
                discountThreshold: { type: number }
                autoAnalyze: { type: boolean }
      responses:
        '200': { description: Updated }
...